Moved tests to test_api and added query parameters
