+++ ohapi/cassettes/test_get_page_with_results.yaml
@@ -7,7 +7,7 @@ interactions:
       Connection: [keep-alive]
       User-Agent: [python-requests/2.9.1]
     method: GET
-    uri: https://www.openhumans.org/api/direct-sharing/project/exchange-member/?access_token=dQ5Phvwaz2kthlKnv0I4ytulMt8fBG
+    uri: https://www.openhumans.org/api/direct-sharing/project/exchange-member/?access_token=ACCESSTOKEN
   response:
     body: {string: '{"created":"created_date_time","project_member_id":"PMI","message_permission":true,"sources_shared":[],"username":"test_user","data":[]}'}
     headers:
